# docs/swagger.yaml
basePath: /api/v1
definitions:
  airquality.AirQuality:
    properties:
      aqi:
        type: integer
      heat_index:
        type: number
      humidity:
        type: number
      pm1_0:
        type: number
      pm2_5:
        type: number
      pm10_0:
        type: number
      temperature:
        type: number
    type: object
  airquality.CurrentAirQuality:
    properties:
      aqi:
        type: integer
      heat_index:
        type: number
      humidity:
        type: number
      last_seen:
        type: string
      location:
        $ref: '#/definitions/airquality.Location'
      pm1_0:
        type: number
      pm2_5:
        type: number
      pm10_0:
        type: number
      sample_count:
        example: 6
        type: integer
      status:
        example: online
        type: string
      temperature:
        type: number
    type: object
  airquality.CurrentResponse:
    properties:
      data:
        $ref: '#/definitions/airquality.CurrentAirQuality'
    type: object
  airquality.CustomRangeResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/airquality.DataPoint'
        type: array
    type: object
  airquality.DataPoint:
    properties:
      label:
        type: string
      metrics:
        $ref: '#/definitions/airquality.AirQuality'
      timestamp:
        type: string
    type: object
  airquality.ErrorResponse:
    properties:
      details:
        example: timeline is required
        type: string
      error:
        example: Missing or invalid query parameters
        type: string
    type: object
  airquality.HistoricalResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/airquality.DataPoint'
        type: array
      timeline:
        example: daily
        type: string
    type: object
  airquality.Location:
    properties:
      lat:
        example: 24.8607
        type: number
      lon:
        example: 67.0011
        type: number
      provider:
        example: mobile
        type: string
      timestamp:
        type: string
    type: object
  airquality.PublicMapDevice:
    properties:
      aqi:
        type: integer
      id:
        type: string
      is_outdoor:
        type: boolean
      lat:
        type: number
      lon:
        type: number
      measured_at:
        type: string
      name:
        type: string
      pm2_5:
        type: number
      status:
        type: string
      temperature:
        type: number
    type: object
  airquality.PublicMapResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/airquality.PublicMapDevice'
        type: array
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Air quality monitoring API
  title: Air Quality API
  version: "1.0"
paths:
  /air-quality/current:
    get:
      description: Returns sensor metrics averaged over the last hour, the device
        status (offline when nothing was received for 20 minutes), the last_seen timestamp,
        and the latest known location. 404 only when the device has never reported.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/airquality.CurrentResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/airquality.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/airquality.ErrorResponse'
      summary: Get current air quality
      tags:
      - Air Quality
  /air-quality/custom:
    get:
      description: Returns air quality data points between start_date and end_date
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start_date
        required: true
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end_date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/airquality.CustomRangeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/airquality.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/airquality.ErrorResponse'
      summary: Get air quality data for a custom date range
      tags:
      - Air Quality
  /air-quality/historical:
    get:
      description: Returns historical air quality data aggregated by timeline
      parameters:
      - description: Timeline
        enum:
        - daily
        - weekly
        - monthly
        - yearly
        in: query
        name: timeline
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/airquality.HistoricalResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/airquality.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/airquality.ErrorResponse'
      summary: Get historical air quality data
      tags:
      - Air Quality
...
  /devices/{id}/current:
    get:
      description: Same as /air-quality/current but restricted to a single device
      parameters:
      - description: Device id (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/airquality.CurrentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/airquality.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/airquality.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/airquality.ErrorResponse'
      summary: Get current air quality for one device
      tags:
      - Air Quality
  /devices/{id}/historical:
    get:
      description: Same as /air-quality/historical but restricted to a single device
      parameters:
      - description: Device id (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Timeline
        enum:
        - daily
        - weekly
        - monthly
        - yearly
        in: query
        name: timeline
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/airquality.HistoricalResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/airquality.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/airquality.ErrorResponse'
      summary: Get historical air quality data for one device
      tags:
      - Air Quality
...
  /public/devices/{id}/current:
    get:
      parameters:
      - description: Device id (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/airquality.CurrentResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/airquality.ErrorResponse'
      summary: Get current readings for a public device
      tags:
      - Public
  /public/devices/{id}/historical:
    get:
      parameters:
      - description: Device id (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Timeline
        enum:
        - daily
        - weekly
        - monthly
        - yearly
        in: query
        name: timeline
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/airquality.HistoricalResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/airquality.ErrorResponse'
      summary: Get historical readings for a public device
      tags:
      - Public
  /public/map/devices:
    get:
      description: Returns the latest complete sample for devices whose owners explicitly
        publish both readings and exact location.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/airquality.PublicMapResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/airquality.ErrorResponse'
      summary: List public devices for the community map
      tags:
      - Public
swagger: "2.0"